Add files via upload
The dataset is called MNIST and refers to handwritten digit recognition. The dataset comprises of 28x28 images of handwritten digits(1 per image) and the goal is to write an algorithm that which digit is written. We will build a network of two hidden layers between inputs and outputs.
In this project i will be using tensorflow as it is the amazing thing when it comes to machine learning. With just few lines of code we can do a lot.